Break up and organise code from index.js
